Fixed a datatype bug in RE assets class, load_RE_profile function

+++ Code/Assets/RE_WIND_Offshore_Lim/__init__.py
@@ -128,12 +128,12 @@ class RE_WIND_Offshore_Lim_Asset(Asset_STEVFNs):
         n_sets = int(np.ceil(self.number_of_edges/set_size))
         gap = int(len(full_profile) / (n_sets * set_size)) * set_size
         offset = set_size * set_number
-        new_profile = np.zeros(n_sets * set_size)
+        new_profile = np.zeros(int(n_sets * set_size))
         for counter1 in range(n_sets):
-            old_loc_0 = offset + gap*counter1
-            old_loc_1 = old_loc_0 + set_size
-            new_loc_0 = set_size * counter1
-            new_loc_1 = new_loc_0 + set_size
+            old_loc_0 = int(offset + gap*counter1)
+            old_loc_1 = int(old_loc_0 + set_size)
+            new_loc_0 = int(set_size * counter1)
+            new_loc_1 = int(new_loc_0 + set_size)
             new_profile[new_loc_0 : new_loc_1] = full_profile[old_loc_0 : old_loc_1]
         self.gen_profile.value = new_profile[:self.number_of_edges]
         return
